Add lib to osc-cost
- create internal lib (for now) for the output
- externalize args, oapi and core
