+++ src/core.rs
@@ -1,12 +1,3 @@
-use crate::ods::ser::to_bytes;
-use crate::prometheus::ser::to_prom;
-use crate::prometheus::ser::CustomLabelKey;
-use comfy_table::modifiers::UTF8_ROUND_CORNERS;
-use comfy_table::presets::UTF8_FULL;
-use comfy_table::Cell;
-use comfy_table::ContentArrangement;
-use comfy_table::Table;
-use log::warn;
 use serde::Deserialize;
 use serde::Serialize;
 use std::collections::HashMap;
@@ -152,150 +143,6 @@ impl Resources {
 
     pub fn cost_per_year(&self) -> Result<f32, ResourceError> {
         Ok(self.cost_per_hour()? * HOURS_PER_MONTH * 12.0)
-    }
-
-    pub fn json(&self) -> serde_json::Result<String> {
-        let mut out = String::new();
-        for resource in &self.resources {
-            match serde_json::to_string(resource) {
-                Ok(serialized) => out.push_str(serialized.as_str()),
-                Err(e) => {
-                    warn!("provide vm serialization: {}", e);
-                    continue;
-                }
-            }
-            out.push('\n');
-        }
-        out.pop();
-        Ok(out)
-    }
-
-    pub fn ods(&self) -> crate::ods::error::Result<Vec<u8>> {
-        to_bytes(&self.resources)
-    }
-    pub fn prometheus(&self) -> crate::prometheus::error::Result<String> {
-        let keep_label = vec![
-            "account_id".to_string(),
-            "osc_cost_version".to_string(),
-            "region".to_string(),
-            "resource_type".to_string(),
-            "resource_id".to_string(),
-            "price_per_hour".to_string(),
-            "price_per_month".to_string(),
-        ];
-
-        let primary_name = "_price_hour".to_string();
-        let primary_help = " price by hour".to_string();
-        let primary_label_key = "price_per_hour".to_string();
-        let secondary_name = "_price_month".to_string();
-        let secondary_help = " price by month".to_string();
-        let secondary_label_key = "price_per_month".to_string();
-        let label_type = "resource_id".to_string();
-        let primary = CustomLabelKey {
-            name: primary_name,
-            help: primary_help,
-            key: primary_label_key,
-        };
-        let secondary = CustomLabelKey {
-            name: secondary_name,
-            help: secondary_help,
-            key: secondary_label_key,
-        };
-        to_prom(&self.resources, keep_label, primary, secondary, label_type)
-    }
-
-    fn get_currency(region: &str) -> String {
-        match region {
-            "eu-west-2" | "cloudgouv-eu-west-1" => String::from("€"),
-            "ap-northeast-1" => String::from("¥"),
-            "us-east-2" | "us-west-1" => String::from("$"),
-            _ => String::from("€"),
-        }
-    }
-
-    pub fn human(&self) -> Result<String, Box<dyn error::Error>> {
-        let mut currency: String = String::new();
-        let mut account_id: String = String::new();
-
-        let mut table_resource = Table::new();
-        table_resource
-            .load_preset(UTF8_FULL)
-            .apply_modifier(UTF8_ROUND_CORNERS)
-            .set_content_arrangement(ContentArrangement::Dynamic)
-            .set_width(100)
-            .set_header(vec![
-                "Resource Type",
-                "Count",
-                "Total price per hour",
-                "Total price per month",
-                "Total price per year",
-            ]);
-
-        for resource in self.resources.iter() {
-            match resource {
-                Resource::Aggregate(agg) => {
-                    if currency.is_empty() {
-                        currency = Resources::get_currency(
-                            agg.region.as_ref().ok_or("could not get the region")?,
-                        );
-                    }
-                    if account_id.is_empty() {
-                        account_id = agg
-                            .account_id
-                            .clone()
-                            .ok_or("could not get the account_id")?;
-                    }
-                    table_resource.add_row(vec![
-                        agg.aggregated_resource_type.clone(),
-                        format!("{}", agg.count),
-                        format!(
-                            "{}{}",
-                            agg.price_per_hour
-                                .ok_or("could not get the price_per_hour")?,
-                            currency
-                        ),
-                        format!(
-                            "{}{}",
-                            agg.price_per_month
-                                .ok_or("could not get the price_per_month")?,
-                            currency
-                        ),
-                        format!(
-                            "{}{}",
-                            agg.price_per_month
-                                .ok_or("could not get the price_per_year")?
-                                * 12.0,
-                            currency
-                        ),
-                    ])
-                }
-                _ => {
-                    warn!("Got a resource which is not an Aggregate for human output");
-                    continue;
-                }
-            };
-        }
-        let mut table = Table::new();
-        table
-            .load_preset(UTF8_FULL)
-            .apply_modifier(UTF8_ROUND_CORNERS)
-            .set_content_arrangement(ContentArrangement::Dynamic)
-            .set_width(100)
-            .add_row(vec![Cell::new("Account Id"), Cell::new(account_id)])
-            .add_row(vec![
-                Cell::new("Total price per hour"),
-                Cell::new(format!("{}{}", self.cost_per_hour()?, currency)),
-            ])
-            .add_row(vec![
-                Cell::new("Total price per month"),
-                Cell::new(format!("{}{}", self.cost_per_month()?, currency)),
-            ])
-            .add_row(vec![
-                Cell::new("Total price per year"),
-                Cell::new(format!("{}{}", self.cost_per_year()?, currency)),
-            ]);
-
-        Ok(format!("Summary:\n{table}\n\nDetails:\n{table_resource}"))
     }
 }
