+++ src/lib.rs
@@ -1,0 +1,4 @@
+const VERSION: &str = env!("CARGO_PKG_VERSION");
+
+pub mod core;
+pub mod oapi;
