+++ src/oapi/vms.rs
@@ -33,9 +33,9 @@ impl Input {
         let result: ReadVmsResponse = loop {
             let filter_vm: FiltersVm = match &self.filters {
                 Some(filter) => FiltersVm {
-                    tag_keys: Some(filter.filter_tag_key.clone()),
-                    tag_values: Some(filter.filter_tag_value.clone()),
-                    tags: Some(filter.filter_tag.clone()),
+                    tag_keys: Some(filter.tag_keys.clone()),
+                    tag_values: Some(filter.tag_values.clone()),
+                    tags: Some(filter.tags.clone()),
                     ..Default::default()
                 },
                 None => FiltersVm::new(),
