--- src/prometheus/mod.rs
@@ -1,2 +1,0 @@
-pub mod error;
-pub mod ser;
